third_party/blink/web_tests/fast/forms/time-multiple-fields/time-multiple-fields-crash-after-adoptnode-expected.txt

Check adoptNode dispatchs blur event for multiple fields time input UI

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS logs.toString() is "start-adoptNode,end-adoptNode,start-focus,end-focus"


PASS successfullyParsed is true

TEST COMPLETE
